fullstackweb/app/retailmenot/requirements.txt
Django==1.6.2
django-statsd-mozilla==0.3.11
supervisor==3.0b2
virtualenv==1.11.4
djangorestframework
django-filter
uwsgi